+++ src/main/java/com/nttdata/masterthesis/javabackend/dao/MappedCategoryDAO.java
@@ -20,55 +20,68 @@ import javax.persistence.TypedQuery;
  * @author MATHAF
  */
 @Stateless
-public class MappedCategoryDAO {
-    
+public class MappedCategoryDAO
+{
+
     @PersistenceContext
     private EntityManager em;
-     
-    public List<MappedCategory> findAll() {
-        TypedQuery<MappedCategory> query = em.createQuery("SELECT trx FROM MappedCategory", MappedCategory.class);
+
+    public List<MappedCategory> findAll()
+    {
+        TypedQuery<MappedCategory> query = em.createQuery( "SELECT trx FROM MappedCategory", MappedCategory.class );
         return query.getResultList();
     }
-  
-    public void save(MappedCategory mappedCategory) {
-        em.persist(mappedCategory);
-    }
-  
-    public void update(MappedCategory mappedCategory) {
-        em.merge(mappedCategory);
-    }
-  
-    public void remove(Long id) {
-        MappedCategory mappedCategory = find(id);
-        if (mappedCategory != null) {
-            em.remove(mappedCategory);
-        }
-    }
-      
-    public void remove(MappedCategory mappedCategory) {
-        if (mappedCategory != null && em.contains(mappedCategory)) {
-            em.remove(mappedCategory);
-        }
-    }
-  
-    public MappedCategory find(Long id) {
-        return em.find(MappedCategory.class, id);
-    }
-     
-    public void detach(MappedCategory mappedCategory) {
-        em.detach(mappedCategory);
+
+    public void save( MappedCategory mappedCategory )
+    {
+        em.persist( mappedCategory );
     }
 
-    public MappedCategory findByMappedId(String mappedId) {
-         Query query = em.createQuery("SELECT c FROM MappedCategory c WHERE c.mappedId = :mappedId");
-         query.setParameter("mappedId", mappedId);
-         try {
-             return (MappedCategory) query.getSingleResult();
-         }catch(NoResultException ex){
-             return null;
-         }
-         
-         
-         
+    public void update( MappedCategory mappedCategory )
+    {
+        em.merge( mappedCategory );
+    }
+
+    public void remove( Long id )
+    {
+        MappedCategory mappedCategory = find( id );
+        if ( mappedCategory != null )
+        {
+            em.remove( mappedCategory );
+        }
+    }
+
+    public void remove( MappedCategory mappedCategory )
+    {
+        if ( mappedCategory != null && em.contains( mappedCategory ) )
+        {
+            em.remove( mappedCategory );
+        }
+    }
+
+    public MappedCategory find( Long id )
+    {
+        return em.find( MappedCategory.class, id );
+    }
+
+    public void detach( MappedCategory mappedCategory )
+    {
+        em.detach( mappedCategory );
+    }
+
+    public MappedCategory findByMappedId( String mappedId )
+    {
+        Query query = em.createQuery( "SELECT c FROM MappedCategory c WHERE c.mappedId = :mappedId" );
+        query.setParameter( "mappedId", mappedId );
+        try
+        {
+            return ( MappedCategory ) query.getSingleResult();
+        } catch ( NoResultException ex )
+        {
+            return null;
+        }
+
+
+
     }
 }
